labs/8laba/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.8
confluent-kafka==2.5.0
kafka-python==2.0.2
fastapi==0.115.0
uvicorn==0.30.6